class RRTrial deleted
