+++ src/main/java/org/skypro/starbank/marketing/component/Invest500Rule.java
@@ -9,28 +9,17 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Component
-public class Invest500Rule implements RecommendationRule {
-    private final RecommendationsRepository repo;
-
-    public Invest500Rule(RecommendationsRepository repo) {
-        this.repo = repo;
+public class Invest500Rule extends SuperclassRecommendationRule {
+    public Invest500Rule(RecommendationsRepository recommendationsRepository) {
+        super(recommendationsRepository);
     }
 
     @Override
-    public Optional<Recommendation> getRecommendation(UUID userId) {
-        Optional<Recommendation> recommendation;
-        if (repo
-                .getSearchResult(userId.toString())
-                .getResult()) {
-            recommendation = Optional.of(new Recommendation(
-                    "Invest 500",
-                    "147f6a0f-3b91-413b-ab99-87f081d60d5a",
-                    "Откройте свой путь к успеху с индивидуальным инвестиционным счетом (ИИС) от нашего банка! Воспользуйтесь налоговыми льготами и начните инвестировать с умом. Пополните счет до конца года и получите выгоду в виде вычета на взнос в следующем налоговом периоде. Не упустите возможность разнообразить свой портфель, снизить риски и следить за актуальными рыночными тенденциями. Откройте ИИС сегодня и станьте ближе к финансовой независимости!"));
-        } else {
-            recommendation = Optional.empty();
-        }
-        return recommendation;
+    public Recommendation getNewRecommendation() {
+        return new Recommendation(
+                "Invest 500",
+                "147f6a0f-3b91-413b-ab99-87f081d60d5a",
+                "Откройте свой путь к успеху с индивидуальным инвестиционным счетом (ИИС) от нашего банка! Воспользуйтесь налоговыми льготами и начните инвестировать с умом. Пополните счет до конца года и получите выгоду в виде вычета на взнос в следующем налоговом периоде. Не упустите возможность разнообразить свой портфель, снизить риски и следить за актуальными рыночными тенденциями. Откройте ИИС сегодня и станьте ближе к финансовой независимости!");
     }
-
 
 }
